Formatted he code and commented

+++ src/main/java/com/drone/drone/dao/MedicationDao.java
@@ -5,8 +5,11 @@ import com.drone.drone.dto.MedicationDto;
 import java.util.List;
 
 public interface MedicationDao {
-    double getLoadedWeightByDroneId (String droneId);
-    double getOpenWeightByMedicationId(List<String> medications);
+    int getLoadedWeightByDroneId(String droneId);
+
+    int getOpenWeightByMedicationId(List<String> medications);
+
     void updateStatus(String medicationId, String droneId);
+
     List<MedicationDto> getMedicationsByDroneId(String droneId);
 }
